chore: Add type stubs for scipy's Rotation class
scipy contains stub files but unfortunately does not export them for use
in other packages.  Therefore create a custom stub file structure and
copy the relevant file from scipy there.

The `def __getattr__(name: str) -> Any: ...` that is added there and in
the `__init__.pyi` files is there so that mypy does not complain about
other scipy types, for which no stubs are provided.

This effort is needed, so that we can use mypy to check if the
rotation is provided in the correct format everywhere.